+++ src/SIP/SipClient.cpp
@@ -1,15 +1,10 @@
 #include "SipClient.hpp"
 
 SipClient::SipClient(std::string number, sockaddr_in address, int expiresSeconds)
-	: _number(std::move(number)), _address(address)
+	: _number(std::move(number)), _address(address),
+	  _expiresSeconds(expiresSeconds),
+	  _expiresAt(std::chrono::steady_clock::now() + std::chrono::seconds(expiresSeconds))
 {
-	renew(expiresSeconds);
-}
-
-void SipClient::renew(int expiresSeconds)
-{
-	_expiresSeconds = expiresSeconds;
-	_expiresAt = std::chrono::steady_clock::now() + std::chrono::seconds(expiresSeconds);
 }
 
 bool SipClient::isExpired(std::chrono::steady_clock::time_point now) const